+++ labs/lab4/Shapes/Canvas.h
@@ -8,7 +8,7 @@ class CCanvas : public ICanvas
 public:
 	void SetColor(Color color) override;
 	void DrawLine(Point const& from, Point const& to) override;
-	void DrawEllipse(Point const& center, size_t horizontalRadius, size_t verticalRadius) override;
+	void DrawEllipse(Point const& center, double horizontalRadius, double verticalRadius) override;
 
 private:
 	Color m_color;
